+++ lib/wrata_api/file_tree.rb
@@ -1,11 +1,14 @@
+# frozen_string_literal: true
+
 module WrataApi
   # Methods to work with file tree
-  module FileTreeMethods
-    def file_tree(project, ref = 'master')
+  module FileListMethods
+    def file_list(project, ref = 'master')
       uri = URI.parse("#{@uri}/runner/file_tree")
       body = {
-          'project' => project,
-          'refs' => ref
+        'flatten' => true,
+        'project' => project,
+        'refs' => ref
       }
       perform_get(uri, body)
     end