+++ spec/wrata_api/wrata_api_file_list_spec.rb
@@ -1,0 +1,11 @@
+# frozen_string_literal: true
+
+require 'spec_helper'
+
+describe WrataApi::WrataApi, '#file_list' do
+  let(:api) { described_class.new }
+
+  it 'file_tree return array' do
+    expect(api.file_list(project)).to be_a(Array)
+  end
+end