+++ Changelog.md
@@ -6,6 +6,7 @@
 
 * Ability to `power_on` servers with different size
 * Change required version of ruby to 2.3
+* Ability to get file list in project
 
 ### Fixes
 * Raise error (not show strange error) while trying to create a lot of server